+++ module/cvm/src/main/clj/convex/db.clj
@@ -1,63 +1,140 @@
 (ns convex.db
 
-  "Etch is a fast, immutable, append-only database specially tailored for cells.
+  "Etch is a fast, immutable, embedded database tailored for cells.
 
-   This namespace provides an API for creating an instance by pointing to a single file. This file
-   hosts an arbitrarily large map of `hash of the encoding of a cell` -> `cell`. Hence, reading require
-   hashes (see [[convex.cell/hash]]) and writes primarilty return refs or nil
-   when not found (see [[convex.ref]]).
+   It can be understood as a data store where keys are hashes of the cells they point to.
+   Hence, the API is pretty simple. [[write]] takes a cell and returns a hash while [[read]]
+   takes a hash and returns a cell (or nil if not found).
 
-   Lastly, a root hash can be stored and retrieved. Cell stored at root is typically used to maintain
-   some sort of global state or table tracking what is needed.
+   Most of the time, usage is even simpler by using [[root-write]] and [[root-read]] to persist
+   the state of a whole application at once (only new values are effectively written).
 
-   **Attention.** By default, R/W functions use the current thread-local database (see [[convex.cvm.db]]).
-   Providing an instance explicitly is tricky because when reading, not all data might be retrieved at once.
-   This is what allows large data, even larger than memory, to be queried: large structures are split into refs
-   and not all refs are necessarily resolved right away. However, any unresolved ref will be resolved against
-   the current thread-local database when needed, not the one that was explicitly provided when reading.
+   Data is retrieved semi-lazily. For instance, in the case of a large vector, only the 
+   the top structure of that vector is fetched. Elements are read from disk when actually accessed
+   then cached using a clever system of soft references under the hood. This explains why data
+   larger than memory can be retrieved and handled since the JVM can garbage-collect those soft
+   references ; their value will be read from disk again if required. Nonetheless, users only deal
+   with cells and all this process is completely transparent.
+
+   Attention, although this namespace is straightforward, one rule must be followed at all time:
+   cells read from an instance can only be written back to that instance. In other words, one
+   must never mix cells read from different instances with the intent of writing them anywhere.
+   This will result in some of the data not being written. Everything, from cells to Etch, has
+   been heavily optimized for Convex peers that only ever handle 1 instance at a time. It is
+   fine using several stores in the same process as long as operations never cross-over.
+
+   Convex tooling, whenever an instance is needed, will always look for the instance associated 
+   with the current thread (if any). The typical workflow is to call [[current-set]] after [[open]]:
+
+   ```clojure
+   (convex.db/current-set (convex.db/open \"my/instance.etch\"))
+   (convex.db/read (convex.db/write (convex.cell/* [:a :b 42])))
+   (convex.db/close)
+   ```
+
+   If no instance is bound to the current thread explicitely, a temporary one is created whenever needed.
+   See [[global-set]] for improving the workflow when an instance is needed in more than one thread.
   
-   In other words, when handling a custom instance, it is best to work on a dedicated thread and call
-   [[convex.cvm.db/local-set]].
-  
-   That being said, instances support multithreading. Being immutable, no thread has to worry that some data might
-   be removed or updated in place."
+   When using a [[convex.cvm/ctx]], its state is initially hold in memory. After opening an Etch
+   instance and setting it as thread-local, this state can be retrieved at any point using [[convex.cvm/state]]
+   and persisted to disk since it is a cell. This renders that state garbage-collecteable as exposed
+   above. Of course, it is important not to close the instance before stopping all operations on that
+   context and its state."
 
   {:author "Adam Helinski"}
 
   (:import (convex.core.data ACell
-                             Hash
-                             Ref)
-           (convex.core.store AStore
-                              MemoryStore)
+                             Hash)
+           (convex.core.store Stores)
            (java.io File)
            (etch EtchStore))
   (:refer-clojure :exclude [flush
-                            read])
-  (:require [convex.cvm.db :as $.cvm.db]
-            [convex.ref    :as $.ref]))
+                            read]))
 
 
 (set! *warn-on-reflection*
       true)
 
 
-(declare read-ref
-         read-root-hash
-         write-ref
-         write-root-ref)
+;;;;;;;;;;
 
 
-;;;;;;;;;; Creating new DBs
+(defn current
+
+  "Returns the thread-local instance (or nil).
+   See [[current-set]]."
+
+  ^EtchStore
+
+  []
+
+  (Stores/current))
+
+
+
+(defn current-set
+
+  "Binds the given `instance` to the current thread.
+   Returns the `instance`.
+   See [[current]]."
+
+  ^EtchStore
+
+  [instance]
+
+  (Stores/setCurrent instance)
+  instance)
+
+
+
+(defn global-set
+
+  "When an instance is used in more than one thread, it is a good idea using this function.
+   Convex tooling will then use the given `instance` in all thread automatically unless it is
+   overwritten with [[current-set]] on a thread per thread basis."
+
+  ^EtchStore
+
+  [instance]
+
+  (Stores/setGlobalStore instance)
+  instance)
+
+
+;;;;;;;;;; Lifecycle
+
+
+(defn close
+
+  "Flushes and closes the thread-local instance. Also unbinds it from the current thread.
+   
+   Note that all instances are also cleanly closed on JVM shutdown but it is
+   more predictable doing it manually."
+
+  []
+
+  (.close (current))
+  (current-set nil)
+  nil)
+
+
+
+(defn flush
+
+  "Flushes the thread-local instance, ensuring all changes are persisted to disk."
+
+  ^EtchStore
+
+  []
+
+  (.flush (current))
+  nil)
+
 
 
 (defn open
 
-  "Opens a db at the given `path`:
-  
-   ```clojure
-   (open \"some_dir/my-db\")
-   ```
-  
+  "Opens an instance at the given `path`.
    File is created if needed."
 
   ^EtchStore
@@ -66,32 +143,16 @@
 
   (let [^File file (File. path)]
     (-> file
-        .getParentFile
-        .mkdirs)
+        (.getParentFile)
+        (.mkdirs))
     (EtchStore/create file)))
 
 
 
-(defn open-in-memory
+(defn open-tmp
 
-  "Alternatively, an in-memory db can be used for some use cases.
-  
-   However, [[open]] is typically recommended."
-
-  ^MemoryStore
-
-  []
-
-  (MemoryStore.))
-
-  
-
-
-(defn open-temp
-
-  "Like [[open]] but works with a temporariy file.
-  
-   A prefix string may be provided."
+  "Like [[open]] but creates a temporary file.
+   A prefix string may be provided for the filename."
 
 
   (^EtchStore []
@@ -102,245 +163,90 @@
   (^EtchStore [prefix]
 
    (EtchStore/createTemp prefix)))
-  
-
-;;;;;;;;;; Lifecycle and miscellaneous
-
-
-(defn close
-
-  "Closes the given `db`."
-
-
-  ([^AStore db]
-
-   (.close db)
-   nil))
-
-
-
-(defn flush
-
-  "Flushes the given `db`, ensuring all changes are persisted to disk.
-  
-   Does not work with in-memory instances."
-
-
-  (^EtchStore []
-
-   (flush ($.cvm.db/local)))
-
-
-  (^EtchStore [^EtchStore db]
-
-   (.flush db)
-   db))
 
 
 
 (defn path
 
-  "Returns the path of the given `db`.
-  
-   Does not work with in-memory instances."
+  "Returns the path of thread-local instance."
+
+  []
+
+  (.getFileName (current)))
 
 
-  ([]
-
-   (path ($.cvm.db/local)))
-
-
-  ([^EtchStore db]
-
-   (.getFileName db)))
-
-
-;;;;;;;;;; Reading
+;;;;;;;;;; R\W
 
 
 (defn read
 
-  "Like [[read-ref]] but ref is directly resolved to its cell.
+  "Reads from the thread-local instance and returns the cell for the given `hash` (or nil
+   if not found)."
   
-   Convenient, but returns nil both when not found or when the stored cell is nil."
-
+  ^ACell
   
-  (^ACell [hash]
+  [^Hash hash]
 
-   (read ($.cvm.db/local)
-         hash))
+  (some-> (.refForHash (current)
+                       hash)
+          (.getValue)))
 
-
-  (^ACell [db hash]
-
-   (some-> ^Ref (read-ref db
-                          hash)
-           .getValue)))
-
-
-
-(defn read-ref
-
-  "Given a `hash`, reads a ref to a cell stored in the `db`, or nil if not found."
-
-
-  (^Ref [hash]
-
-   (read-ref ($.cvm.db/local)
-             hash))
-
-
-  (^Ref [^AStore db ^Hash hash]
-
-   (.refForHash db
-                hash)))
-
-
-
-(defn read-root
-
-  "Like [[read-root-ref]] but ref is directly resolved to its cell, akin to [[read]]."
-
-
-  (^ACell []
-
-   (read-root ($.cvm.db/local)))
-
-
-  (^ACell [db]
-
-   (read db
-         (read-root-hash db))))
-
-
-
-(defn read-root-hash
-
-  "Reads the hash stored as root."
-
-
-  (^Hash []
-
-   (read-root-hash ($.cvm.db/local)))
-
-
-  (^Hash [^AStore db]
-
-   (.getRootHash db)))
-
-
-
-(defn read-root-ref
-
-  "Like [[read-ref]] but uses the hash from [[read-root-hash]]."
-
-
-  (^Ref []
-
-   (read-root-ref ($.cvm.db/local)))
-
-
-  (^Ref [db]
-
-   (read-ref db
-             (read-root-hash db))))
-
-
-;;;;;;;;;; Writing
 
 
 (defn write
 
-  "Wraps `cell` in a ref and calls [[write-ref]]."
+  "Writes the given `cell` to the thread-local instance and returns its hash.
+
+   Very basic cells are not persisted because that would be inefficient and hardly ever happens.
+   They are typically embedded in collections. Hence, this function will return nil for:
+
+     - Address
+     - Empty collections
+     - Primitives (boolean, byte, double, long)
+     - Symbolic (keywords and symbols)"
+
+  ^Hash
+
+  [^ACell cell]
+   
+  (-> (ACell/createPersisted cell)
+      (.cachedHash)))
 
 
-  (^Ref [^ACell cell]
-
-    ;; TODO. Cannot user `ACell.createPersisted()` because of https://github.com/Convex-Dev/convex/issues/298
-    ;;
-    (write ($.cvm.db/local)
-           cell))
+;;;
 
 
-  (^Ref [db ^ACell cell]
+(defn root-read
 
-   (write-ref db
-              ($.ref/create-direct cell))))
+  "Returns the cell stored at the root of the thread-local instance.
 
+   The root is a place in the instance that can be read without providing a hash. It is commonly
+   used for storing the whole state of an application or at least some sort of index containing
+   hashes of other data in the instance. This makes Etch self-sufficient as no hash must be stored
+   externally.
 
+   See [[write-root]]."
 
-(defn write-ref
+  ^ACell
 
-  "Given a ref, writes its associated cell."
+  []
 
-
-  (^Ref [ref]
-
-   (write-ref ($.cvm.db/local)
-              ref))
-
-
-  (^Ref [^AStore db ^Ref ref]
-
-   (.storeTopRef db
-                 ref
-                 Ref/PERSISTED
-                 nil)))
-
-
-
-(defn write-root
-
-  "Wraps `cell` in a ref and calls [[write-root-ref]]."
-
-
-  (^Ref [cell]
-
-   (write-root ($.cvm.db/local)
-               cell))
-
-
-  (^Ref [db ^ACell cell]
-
-   (write-root-ref db
-                   ($.ref/create-direct cell))))
+  (read (.getRootHash (current))))
 
 
 
-(defn write-root-hash
+(defn root-write
 
-  "Writes the given `hash` as root."
+  "Writes the given `cell` to the root of the thread-local instance and returns its hash.
+   Behaves like [[write]].
 
+   See [[read-root]]."
 
-  (^Hash [hash]
+  ^Hash
 
-   (write-root-hash ($.cvm.db/local)
-                    hash))
+  [^ACell cell]
 
-
-  (^Hash [^AStore db hash]
-
-   (.setRootHash db
-                 hash)
-   hash))
-
-  
-
-(defn write-root-ref
-
-  "Writes `ref` and then sets its hash as root."
-
-
-  (^Ref [ref]
-
-   (write-root-ref ($.cvm.db/local)
-                   ref))
-
-
-  (^Ref [db ref]
-
-   (let [^Ref ref-2 (write-ref db
-                               ref)]
-     (write-root-hash db
-                      (.cachedHash ref-2))
-     ref-2)))
+  (let [hash (write cell)]
+    (.setRootHash (current)
+                  hash)
+    hash))