+++ module/cvm/src/test/clj/convex/test/cell.clj
@@ -68,107 +68,109 @@
 (T/deftest *
 
   (T/is (= ($.read/string "[:a #42]")
-           ($.cell/* [:a ~($.cell/address 42)]))
+           ($.cell/* ([:a ~($.cell/address 42)])))
         "Unquoting leaves objects as is")
 
   (T/is (= ($.read/string ":foo")
-           ($.cell/* :foo))
+           ($.cell/* (:foo)))
         "Keyword")
 
   (T/is (= ($.read/string "(conj [2] :foo)")
-           ($.cell/* (conj [2]
-                           :foo)))
+           ($.cell/* ((conj [2]
+                            :foo))))
         "List")
 
   (T/is (= ($.read/string "{:a :b}")
-           ($.cell/* {:a :b}))
+           ($.cell/* ({:a :b})))
         "Map")
 
   (T/is (= ($.read/string "#{:a :b}")
-           ($.cell/* #{:a :b}))
+           ($.cell/* (#{:a :b})))
         "Set")
 
   (T/is (= ($.read/string "[:a :b]")
-           ($.cell/* [:a :b]))
+           ($.cell/* ([:a :b])))
         "Vector")
 
   (T/is (= ($.read/string "foo")
-           ($.cell/* foo))
+           ($.cell/* (foo)))
         "Symbol")
 
   (T/is (= ($.read/string "true")
-           ($.cell/* true))
+           ($.cell/* (true)))
         "Boolean")
 
   (T/is (= ($.read/string "\\c")
-           ($.cell/* \c))
+           ($.cell/* (\c)))
         "Char")
 
   (T/is (= ($.read/string "42.42")
-           ($.cell/* 42.42))
+           ($.cell/* (42.42)))
         "Double")
 
   (T/is (= ($.read/string "42")
-           ($.cell/* 42))
+           ($.cell/* (42)))
         "Long")
 
   (T/is (= ($.read/string "\"foo\"")
-           ($.cell/* "foo"))
+           ($.cell/* ("foo")))
         "String"))
 
 
 
 (T/deftest any
 
-  (T/is (= ($.read/string "#42")
+  ;; Do not use `$.cell/*`.
+
+  (T/is (= (first ($.read/string "#42"))
            ($.cell/any ($.cell/address 42)))
         "By default, returns object as is")
 
-  (T/is (= ($.read/string "(+ 2 :foo)")
+  (T/is (= (first ($.read/string "(+ 2 :foo)"))
            ($.cell/any (cons '+
                              [2 :foo])))
         "Seq")
 
-  (T/is (= ($.read/string ":foo")
+  (T/is (= (first ($.read/string ":foo"))
            ($.cell/any :foo))
         "Keyword")
 
-  (T/is (= ($.read/string "(+ 2 :foo)")
+  (T/is (= (first ($.read/string "(+ 2 :foo)"))
            ($.cell/any '(+ 2 :foo)))
         "List")
 
-  (T/is (= ($.read/string "{:a :b}")
+  (T/is (= (first ($.read/string "{:a :b}"))
            ($.cell/any {:a :b}))
         "Map")
 
-  (T/is (= ($.read/string "#{:a :b}")
+  (T/is (= (first ($.read/string "#{:a :b}"))
            ($.cell/any #{:a :b}))
         "Set")
 
-  (T/is (= ($.read/string "[:a :b]")
+  (T/is (= (first ($.read/string "[:a :b]"))
            ($.cell/any [:a :b]))
         "Vector")
 
-  (T/is (= ($.read/string "foo")
+  (T/is (= (first ($.read/string "foo"))
            ($.cell/any 'foo))
         "Symbol")
 
-  (T/is (= ($.read/string "true")
+  (T/is (= (first ($.read/string "true"))
            ($.cell/any true))
         "Boolean")
 
-  (T/is (= ($.read/string "\\c")
+  (T/is (= (first ($.read/string "\\c"))
            ($.cell/any \c))
         "Char")
 
-  (T/is (= ($.read/string "42.42")
+  (T/is (= (first ($.read/string "42.42"))
            ($.cell/any 42.42))
         "Double")
 
-  (T/is (= ($.read/string "42")
+  (T/is (= (first ($.read/string "42"))
            ($.cell/any 42))
         "Long")
 
-  (T/is (= ($.read/string "\"foo\"")
+  (T/is (= (first ($.read/string "\"foo\""))
            ($.cell/any "foo"))
         "String"))